feat: refactor ProviderIndicator into LlmSwitcher and LlmIndicator
### UI & Components
- Replaced the monolithic `ProviderIndicator` with two specialized components: `LlmSwitcher` (interactive) and `LlmIndicator` (read-only).
- Refactored internal display logic into private `ProviderIcon` and `ModelLabel` helper functions for better maintainability.
- Updated `ModelLabel` shortening logic to include `mistral` in the list of prefixes to be stripped.
- Migrated `SummaryCard` to use `LlmIndicator` and `HomeScreen` to use `LlmSwitcher`.
- Refined the layout of model labels, including improved positioning logic based on box size.
